// bksharing-be-main/src/shared/enums/common.enum.ts
export enum SYMBOL {
  PLUS = '+',
  MINUS = '-',
  TIMES = '*',
  DIVIDE = '/',
  COMMA = ',',
  DOT = '.',
}

export enum DateRangeFilter {
  ALL = 'ALL',
  TODAY = 'TODAY',
  ONE_WEEK = '1_WEEK',
  ONE_MONTH = '1_MONTH',
  THREE_MONTHS = '3_MONTHS',
  SIX_MONTHS = '6_MONTHS',
  ONE_YEAR = '1_YEAR',
}